test: pass worker count by MAKEFLAGS -j argument
Modified tests/Makefile to pass worker count by MAKEFLAGS -j argument.
Work around of (probably) a bug in make 4.1 not passing -j N to the
MAKEFLAGS variable.

If the N is missing for -j option, assume it uses full cpu resources.
The manual page of GNU make says it has no limit in such case.

If the -j option itself is missing, it'll run tests in a single task
like it did before.

before:
    $ make test TESTARG=-j2

after:
    $ make test -j 2
    ...
    ./runtest.py -j2
    Start 211 tests with 2 worker

Signed-off-by: Daniel T. Lee <danieltimlee@gmail.com>
Signed-off-by: Namhyung Kim <namhyung@gmail.com>